Styling
I accidentally went and did all of it. I don’t think it’s *quite* up to
my usual high standard – but I was a bit pressed for time… sorry!

Obviously it’s fine if you want to change it or not use it at all.

For reference this is where I got the colour scheme:

http://dribbble.com/shots/1470913-Seattle?list=popular&offset=75